+++ CHANGELOG.md
@@ -10,6 +10,20 @@
 
 ##### Bug Fixes
 
+- None.
+
+## 3.5.1 (2026-02-15)
+
+##### Breaking
+
+- None.
+
+##### Enhancements
+
+- None.
+
+##### Bug Fixes
+
 - Fix false positive redundant public accessibility for types used in same-type generic requirements on extensions, e.g. `extension SomeProtocol where Self == MyType`.
 - Exclude declarations ignored by a file-level `// periphery:ignore:all` comment from superfluous ignore comment detection.
 - Superfluous ignore comment results are now included in baselines.